Add LoopedDecoder
LoopedDecoder works like the normal Decoder, but when the decoder
reaches the end of file it will restart playback from the beginning.
This has the advantage over Decoder::new().repeat_infinite() that it
will not buffer the entire uncompressed audio file into memory.

Update minimp3 to version 0.5 to use the `into_iter()` method