Make `DynamicMixer` and `OutputStream` generic
This commit adds a new trait, `Mixer`, and refactors some of the
code around generic mixers in order to support it. This is to
make more of the code reusable without rewriting - specifically,
to support a feature that I would like to implement in
`bevy_audio` where audio sources can send audio to an
intermediate mixer. The motivation for this was to add a global
limiter to all game audio, so that loud sounds did not cause
clipping on the master bus, but another possible usecase would
be to split in-game audio and menu audio into separate "worlds"
and apply, for example, a low-pass filter while in the pause
screen.