The cpal version requirement is too low causing #573
Specifically cpal changed `SupportedStreamConfigRange` and made it Copy.
As that relaxes the requirements on it its a non breaking change. Rodio was
then refactored to remove the now no longer needed `.clone()` in 77c5c54. That
clone however is still needed for Rodio users with a lower cpal version. This
commit fixes that by requiring a higher cpal version.