Adds split_once to Source
split_once can be used chop a source into two sources without
duplicating the first. Those sources can then not be played simultaneously.
This is usefull for adding a fadeout at the end of a song for example.
Though we could later specialize fadeout at end.

The implementation uses a `SplitAt` struct, `split_once` returns an array of
two of those. In the future we can introduce methods like `split` that
return a vec of SplitAt's allowing users to chop a source into multiple
sections