This makes ChannelCount NonZero<u16> and channels not zero asserts
I ran into a lot of bugs while adding tests that had to do with channel
being set to zero somewhere. While this change makes the API slightly
less easy to use it prevents very hard to debug crashes/underflows etc.

Performance might drop in decoders, the current implementation makes the
bound check every time `channels` is called which is once per span. This
could be cached to alleviate that.