Only close cpal::Stream once all mixers/sinks drop
This resolves a common footgun in rodio, that is: accidentily
dropping the stream handle causing all audio to stop. This can be hard
to debug.

To solve that we make `cpal::Stream` undroppable by the user. Instead we
drop it for the user once all mixers/sinks that are connected to the
`cpal::Stream` are dropped. The only way to make `cpal::Stream`
undroppable is to hold it in a sleeping thread. This is because
`cpal::Stream` is neither send nor sync. If it where we could just store
it on the heap in an `Arc` shared by all sinks and streams.

The `cpal::Stream` is now created on a seperate thread. After creation
that thread sleeps until the OutputStreamTracker reaches a count of
zero. The OutputStreamTracker is shared between all mixers and sinks
that are connected to a `cpal::Stream`.

This design is not perfect, its still quite easy to drop the
`cpal::Stream`. Simply take a normal mixer (not the `OutputMixer`
introduced by this PR) and connect it to an `OutputMixer`. Then use the
normal mixer instead of the `OutputMixer` and drop the `OutputMixer`.

An alternative, far simpler solution could be a drop bomb