Print/log when dropping the OutputStream
This was a real footgun for new users, they would drop the OutputStream
and be suprised their audio stopped playing. The message explicitly
mentions dropping stops the audio.

We print on stderr unless the tracing is enabled. If for some reason a
user needs a clean stderr and does not want to use tracing they can
disable the print using `stream.log_on_drop(false)`.

We do not print if the stream is dropped because the program is
panicking. Any print there could confuse the user as to the cause of the
panic. It could also interleave with the panic output.

This replaces:
- https://github.com/RustAudio/rodio/pull/746
- https://github.com/RustAudio/rodio/pull/747