Draft of an optimized pausing architecture
`Source`s in rodio get queried by the `OutputStream`. That means that the
stream calls the next method on a `Source`. Only then can the `Source` do
something, like call next on a `Source` it wraps.

For pausing we want to turn this flow around, we want to affect the
`OutputStream` from a `Source` (possibly wrapped by many others).

To do so we need to get access to the `OutputStream` in a `Source`. We
do so by making the `OutputStream` call `set_pause_handle(<some handle>)`
on each `Source` added to it. During that call of `set_pause_handle`
each of those sources calls `set_pause_handle` on which ever source the
wrap. This goes on all through the `Source`s tree.

A `Pausable` source keeps a copy of the pause handle passed to it when
`set_pause_handle` got called on it. When it needs to pause the stream
it can simply call `pause()` on its stored pause handle.

There will be multiple types of pause handles:
- One which simply (un)pauses the `cpal::Stream`. It is passed to all
  the sources through `set_pause_handle` on `OutputStream` creation.
- One made for `Mixer`. It is passed to all streams added to the `Mixer`.
  This wraps whichever pause handle was set on the mixer. It only calls
  pause on that handle when all streams in the mixer have called pause
  on it.
- Maybe more I did not think off?