Adds a builder for configuring outputstreams
Its basically the same as the microphone builder. It will replace the
OutputStream. In the future we'll add more abstractions on top. Until
its done it lives under an experimental flag.

Names are subject to change too, Speakers is probably not ideal but it
conveys the meaning better then OutputStream. I'm thinking of having a
Source -> Stream -> Sink terminolgy where a Sink could be the audio
card, the network or a file (the wavwriter).